# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Edna Bar Book Printing API
  description: |
    A comprehensive RESTful API for managing book printing operations.
    
    This API provides complete functionality for:
    - Managing books, authors, and publishers
    - Handling printing companies (private and contracted)
    - Creating and tracking printing contracts
    - Scheduling and monitoring printing jobs
    - Real-time delivery tracking
    
    Built with Go following clean architecture principles.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/edna-bar/api
    email: support@edna-bar.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://edna-bar.com/terms
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.edna-bar.com/api
    description: Production server
...
paths:
...
  /livros:
    get:
      tags:
        - books
      summary: List all books
      description: Retrieve a list of all books in the system
      parameters:
        - name: limit
          in: query
          description: Maximum number of books to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of books to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: title
          in: query
          description: Filter books by title (partial match)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - books
      summary: Create a new book
      description: Add a new book to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookRequest'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /livros/{isbn}:
    get:
      tags:
        - books
      summary: Get book by ISBN
      description: Retrieve a specific book by its ISBN
      parameters:
        - name: isbn
          in: path
          required: true
          description: Book ISBN
          schema:
            type: string
            pattern: '^(97[89])?[\d\-]{10,17}$'
      responses:
        '200':
          description: Book found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - books
      summary: Update book
      description: Update an existing book
      parameters:
        - name: isbn
          in: path
          required: true
          description: Book ISBN
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookRequest'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - books
      summary: Delete book
      description: Remove a book from the system
      parameters:
        - name: isbn
          in: path
          required: true
          description: Book ISBN
          schema:
            type: string
      responses:
        '204':
          description: Book deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /livros/{isbn}/authors:
    get:
      tags:
        - books
      summary: Get book authors
      description: Retrieve all authors of a specific book
      parameters:
        - name: isbn
          in: path
          required: true
          description: Book ISBN
          schema:
            type: string
      responses:
        '200':
          description: Authors found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - books
      summary: Add author to book
      description: Associate an author with a book
      parameters:
        - name: isbn
          in: path
          required: true
          description: Book ISBN
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rg:
                  type: string
                  description: Author's RG
              required:
                - rg
      responses:
        '201':
          description: Author added to book successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Authors endpoints
...
components:
  schemas:
    Book:
      type: object
      properties:
        isbn:
          type: string
          description: Book ISBN
          example: "978-0307474728"
        titulo:
          type: string
          description: Book title
          example: "Cem Anos de Solidão"
        data_de_publicacao:
          type: string
          format: date
          description: Publication date
          example: "1967-06-05"
        editora_id:
          type: integer
          description: Publisher ID
          example: 1
      required:
        - isbn
        - titulo
        - data_de_publicacao
        - editora_id
    CreateBookRequest:
      type: object
      properties:
        isbn:
          type: string
          description: Book ISBN
          pattern: '^(97[89])?[\d\-]{10,17}$'
          example: "978-0307474728"
        titulo:
          type: string
          description: Book title
          maxLength: 255
          example: "Cem Anos de Solidão"
        data_de_publicacao:
          type: string
          format: date
          description: Publication date
          example: "1967-06-05"
        editora_id:
          type: integer
          description: Publisher ID
          minimum: 1
          example: 1
      required:
        - isbn
        - titulo
        - data_de_publicacao
        - editora_id
    UpdateBookRequest:
      type: object
      properties:
        titulo:
          type: string
          description: Book title
          maxLength: 255
          example: "Cem Anos de Solidão - Edição Especial"
        data_de_publicacao:
          type: string
          format: date
          description: Publication date
          example: "1967-06-05"
        editora_id:
          type: integer
          description: Publisher ID
          minimum: 1
          example: 1
    Author:
      type: object
      properties:
        rg:
          type: string
          description: Author RG
          example: "12345678901"
        nome:
          type: string
          description: Author name
          example: "Gabriel García Márquez"
        endereco:
          type: string
          description: Author address
          example: "América Latina"
      required:
        - rg
        - nome
        - endereco
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
        code:
          type: string
          description: Error code
          example: "NOT_FOUND"
        details:
          type: string
          description: Additional error details
          example: "Book with ISBN 978-0123456789 was not found"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
      required:
        - error
        - code
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid input data"
            code: "BAD_REQUEST"
            details: "ISBN must be a valid format"
            timestamp: "2024-01-15T10:30:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
            details: "Book with ISBN 978-0123456789 was not found"
            timestamp: "2024-01-15T10:30:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            details: "An unexpected error occurred"
            timestamp: "2024-01-15T10:30:00Z"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# Global security (can be overridden per endpoint)
# security:
#   - bearerAuth: []